refactor: split interactionAccessibility spec into component/hook unit tests
