+++ src/shared/ui/feedback/Overlay.spec.tsx
@@ -52,4 +52,17 @@ describe("shared overlay surface", () => {
     fireEvent.click(overlay);
     expect(onClick).toHaveBeenCalledOnce();
   });
+
+  it("activates Overlay with Enter", () => {
+    const onClick = vi.fn();
+    render(
+      <Overlay position="center" onClick={onClick}>
+        Close
+      </Overlay>
+    );
+
+    fireEvent.keyDown(screen.getByRole("button", { name: "Close" }), { key: "Enter" });
+
+    expect(onClick).toHaveBeenCalledOnce();
+  });
 });
